Delete common Data + Propulsion repo
calculate common data in one place.
set propulsion repository
