+++ last_letter_2/config/param/aircrafts/HCUAV/aerodynamics.yaml
