+++ last_letter_2/config/param/aircrafts/aerosonde/aerodynamics.yaml
