+++ last_letter_2/config/param/aircrafts/firefly_Y6/aerodynamics.yaml
