--- last_letter_2/config/param/propulsion.yaml
@@ -1,10 +1,0 @@
-#type 0 = No engine, 1= beardEngine
-motor1/motorType: 1
-
-#motor Engine Parameters
-motor1/s_prop: 0.2027
-motor1/k_motor: 80
-motor1/k_t_p: 0.0
-motor1/k_omega: 20.0
-motor1/c_prop: 1.0
-
