+++ last_letter_2/include/propulsion/noEngine.hpp
@@ -1,0 +1,8 @@
+
+class NoEngine : public Propulsion
+ {
+   public:
+   NoEngine(Model *);
+   void calcThrust();
+   void calcTorque();
+ };
