+++ last_letter_2/launch/simulation.launch
@@ -27,11 +27,11 @@
         args="-z 1.0 -Y 2.1  -urdf -model plane -param robot_description" respawn="false" output="screen" /> 
           
   <rosparam command="load" file="$(find last_letter_2)/config/param/HID.yaml" />
-  <rosparam command="load" file="$(find last_letter_2)/config/param/HCUAV/aerodynamics.yaml" />
+  <rosparam command="load" file="$(find last_letter_2)/config/param/aircrafts/HCUAV/aerodynamics.yaml" />
   <rosparam command="load" file="$(find last_letter_2)/config/param/environment.yaml" />
   <rosparam command="load" file="$(find last_letter_2)/config/param/world.yaml" />
   <rosparam command="load" file="$(find last_letter_2)/config/param/simulation.yaml" />
-  <rosparam command="load" file="$(find last_letter_2)/config/param/propulsion.yaml" />
+  <rosparam command="load" file="$(find last_letter_2)/config/param/propulsion/beardEngine.yaml" />
   
   <node pkg="joy" type="joy_node" name="joystick" output="screen" >
     <param name="dev" value="/dev/input/js0" />
