+++ last_letter_2/src/propulsion/propulsion.cpp
@@ -1,0 +1,14 @@
+
+#include "noEngine.cpp"
+#include "beardEngine.cpp"
+
+Propulsion::Propulsion(Model *parent)
+{
+    model=parent;
+}
+
+void Propulsion::calcWrench()
+{
+    calcThrust();
+    calcTorque();
+}
